api build up

+++ sys/handle/user.js
@@ -1,34 +1,47 @@
-const { user } = require("./sql.js").models;
+const { user } = require("./model.js").models;
 const crypto = require("crypto");
 const state = require("../config/state.js");
-
-const emailReg = /^([a-zA-Z0-9_-])+@([a-zA-Z0-9_-])+((\.[a-zA-Z0-9_-]{2,3}){1,2})$/;
-const isEmpty = (string) => string === undefined ? true : string == null ? true : string === "";
+const { isEmail, isEmpty, isNumber } = require("./utils.js");
 
 const handle = {
+    getUser: async data => {
+        if (isEmpty(data)) return state.Empty;
+        return isNumber(data) ? await handle.getUserById(data) : await handle.getUserNameOrEmail(data);
+    },
     /**
      * @param {number} id
-     * @return {any|null}
+     * @return {user|state}
      */
     getUserById: async id => {
         const result = await user.findByPk(id);
-        return result ? result : null;
+        return result ? result : state.NotExist;
     },
     /**
      * @param {string} name
-     * @return {any|null}
+     * @return {user|state}
      */
     getUserByName: async name => {
         const result = await user.findOne({ where: { name: name }});
-        return result ? result : null;
+        return result ? result : state.NotExist;
     },
     /**
      * @param {string} email
-     * @return {any|null}
+     * @return {user|state}
      */
     getUserByEmail: async email => {
         const result = await user.findOne({ where: { email: email }});
-        return result ? result : null;
+        return result ? result : state.NotExist;
+    },
+    /**
+     * @param {string} data
+     * @return {user|state}
+     */
+    getUserNameOrEmail: async data => {
+        if (isEmail(data) && data.length <= 64)
+            return await handle.getUserByEmail(data);
+        else if (data.length <= 16)
+            return await handle.getUserByName(data);
+        else return state.OverSize;
     },
     /** @return {number} */
     getLastId: async () => {
@@ -38,42 +51,46 @@ const handle = {
      * @param {number} [num]
      * @return any[]
      */
-    getAllUsers: async (num) => {
-        return num ? await user.findAll({ limit: num }) : await user.findAll();
+    getAllUsers: async num => {
+        if (num) {
+            if (!isNumber(num)) return state.Empty;
+            return await user.findAll({ limit: num });
+        } else {
+            return await user.findAll();
+        }
     },
     /**
      * login by email or name
-     * @param {{account:string,password:string}} json
-     * @return {any|number|null}
+     * @param {string} account
+     * @param {string} password
+     * @return {any|null}
      */
-    login: async json => {
-        let account = json.account, password = json.password, user;
+    login: async (account, password) => {
+        let tempPsw, user;
         if (isEmpty(account) || isEmpty(password)) return state.Empty;
-        password = crypto.createHash("md5").update(account + password).digest("hex");
-        if (emailReg.test(account) && account.length <= 64)
-            user = await handle.getUserByEmail(account);
-        else if (account.length <= 16)
-            user = await handle.getUserByName(account)
-        else return state.OverSize;
-        return user != null ? user.password === password ? user : state.NotCorrect : state.NotExist;
+        tempPsw = crypto.createHash("md5").update(account + password).digest("hex");
+        user = await handle.getUserNameOrEmail(account);
+        return user.password ? user.password === tempPsw ? user : state.NotCorrect : user;
     },
     /**
      * Register
-     * @param {{account:string,password:string}} json
-     * @return {any|number|null}
+     * @param {string} account
+     * @param {string} password
+     * @return {any|null}
      */
-    register: async json => {
-        let account = json.account, password = json.password;
+    register: async (account, password) => {
         if (isEmpty(account) || isEmpty(password)) return state.Empty;
-        if (emailReg.test(account) && (account.length <= 64))
-            return await user.create({ email: account, password: password });
-        else if (account.length <= 16)
-            return await user.create({ name: account, password: password });
-        else return state.OverSize;
+        if (isEmail(account) && (account.length <= 64)) {
+            if ((await handle.getUserByName(account)).id) return state.Duplicate;
+            return await user.create({email: account, password: password});
+        } else if (account.length <= 16) {
+            if ((await handle.getUserByName(account)).id) return state.Duplicate;
+            return await user.create({name: account, password: password});
+        } else return state.OverSize;
     },
     /**
      * Update user mysql
-     * @param {number} userId id
+     * @param {number} id id
      * @param partial data
      * @param [partial.id] id
      * @param [partial.name] name
@@ -82,13 +99,24 @@ const handle = {
      * @param [partial.password] password
      * @param [partial.introduction] introduction
      */
-    update: async (userId, partial) => {
-        return await user.update(partial, { where: { id: userId }});
+    updateById: async (id, partial) => {
+        if (!isNumber(id)) return state.Empty;
+        if (partial.id) delete partial.id;
+        return await user.update(partial, { where: { id: id }});
     },
-    abort: async id => {
+    uploadImageById: async (id, image) => {
+        if (!isNumber(id)) return state.Empty;
+    },
+    deprecateById: async id => {
+        if (!isNumber(id)) return state.Empty;
         const result = await user.destroy({ where: { id: id }});
-        return result > 0 ?  result : null
+        return Number(result) > 0 ?  result : state.NotExist;
+    },
+    restoreById: async id => {
+        if (!isNumber(id)) return state.Empty;
+        const result = await user.restore({ where: { id: id }});
+        return Number(result) > 0 ? result : state.NotCorrect;
     }
-}
+};
 
 module.exports = handle;
